pyproject.toml
[tool.poetry]
name = "linebot"
version = "0.1.0"
description = ""
authors = ["Kevin Liu <kevinxo328@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
load-dotenv = "^0.1.0"
line-bot-sdk = "^2.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
